Limit transport_sockets source files to one logger ID
