+++ test/extensions/transport_sockets/starttls/starttls_socket_test.cc
@@ -7,6 +7,7 @@
 
 #include "source/common/network/transport_socket_options_impl.h"
 #include "source/extensions/transport_sockets/starttls/starttls_socket.h"
+#include "source/extensions/transport_sockets/starttls/starttls_socket_factory.h"
 
 #include "test/mocks/network/mocks.h"
 #include "test/mocks/network/transport_socket.h"
