+++ test/extensions/transport_sockets/starttls/upstream_starttls_integration_test.cc
@@ -1,3 +1,4 @@
+#include "envoy/extensions/transport_sockets/starttls/v3/starttls.pb.h"
 #include "envoy/network/filter.h"
 #include "envoy/server/filter_config.h"
